Moved client initialization into AbstractIndexHandler
