+++ src/main/java/fi/otavanopisto/kuntaapi/server/index/AbstractIndexHander.java
@@ -5,6 +5,8 @@ import java.net.UnknownHostException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 import javax.inject.Inject;
 
 import org.apache.commons.lang3.StringUtils;
@@ -17,12 +19,59 @@ import org.elasticsearch.transport.client.PreBuiltTransportClient;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import fi.otavanopisto.kuntaapi.server.integrations.KuntaApiConsts;
+import fi.otavanopisto.kuntaapi.server.settings.SystemSettingController;
+
 public abstract class AbstractIndexHander {
+
+  private static final String DEFAULT_INDEX = "kunta-api";
+  private static final String DEFAULT_CLUSTERNAME = "elasticsearch";
+  private static final String[] DEFAULT_HOSTS = new String[] {
+    "localhost:9300"
+  };
+
+  @Inject
+  private SystemSettingController systemSettingController;
   
   @Inject
   private Logger logger;
 
-  protected TransportClient createClient(String clusterName, String[] hosts) {
+  private String index;
+  private TransportClient client;
+  
+  @PostConstruct
+  public void init() {
+    String[] hosts = systemSettingController.getSettingValues(KuntaApiConsts.SYSTEM_SETTING_ELASTIC_SEARCH_HOSTS, DEFAULT_HOSTS);
+    String clusterName = systemSettingController.getSettingValue(KuntaApiConsts.SYSTEM_SETTING_ELASTIC_CLUSTER_NAME, DEFAULT_CLUSTERNAME);
+    index = systemSettingController.getSettingValue(KuntaApiConsts.SYSTEM_SETTING_ELASTIC_INDEX, DEFAULT_INDEX);
+    client = createClient(hosts, clusterName);
+  }
+  
+  @PreDestroy
+  public void deinit() {
+    closeClient(client);
+  }
+  
+  protected TransportClient getClient() {
+    return client;
+  }
+  
+  protected byte[] serialize(Indexable indexable) {
+    ObjectMapper objectMapper = new ObjectMapper();
+    try {
+      return objectMapper.writeValueAsBytes(indexable);
+    } catch (JsonProcessingException e) {
+      logger.log(Level.SEVERE, "Failed to serialize indexable object", e);
+    }
+    
+    return new byte[0];
+  }
+  
+  protected String getIndex() {
+    return index;
+  }
+  
+  private TransportClient createClient(String[] hosts, String clusterName) {
     TransportClient transportClient = null;
     
     try {
@@ -53,22 +102,9 @@ public abstract class AbstractIndexHander {
     return transportClient;
   }
   
-  protected void closeClient(TransportClient transportClient) {
+  private void closeClient(TransportClient transportClient) {
     transportClient.close();
   }
-  
-  protected byte[] serialize(Indexable indexable) {
-    ObjectMapper objectMapper = new ObjectMapper();
-    try {
-      return objectMapper.writeValueAsBytes(indexable);
-    } catch (JsonProcessingException e) {
-      logger.log(Level.SEVERE, "Failed to serialize indexable object", e);
-    }
-    
-    return new byte[0];
-  }
-  
-  public abstract String getIndex();
   
   private void prepareIndex(TransportClient transportClient) {
     if (!indexExists(transportClient)) {
